+++ src/providers/index.js
@@ -33,6 +33,7 @@ module.exports = [
   require('./sentry'),
   require('./snapengage'),
   require('./usercycle'),
+  require('./userfox'),
   require('./uservoice'),
   require('./vero'),
   require('./woopra')