+++ History.md
@@ -1,3 +1,39 @@
+0.9.17 / 2013-5-1
+=================
+* Updated Keen to version 2.1.0
+
+0.9.16 / 2013-4-30
+==================
+* Fixed bug affecting Pingdom users
+
+0.9.15 / 2013-4-30
+==================
+* Added identify to UserVoice
+
+0.9.14 / 2013-4-29
+==================
+* Fixing userfox integration to accept all traits not just signup_date
+
+0.9.13 / 2013-4-29
+==================
+* Fixing ordering of ignore referrer option in Google Analytics
+
+0.9.12 / 2013-4-27
+==================
+* Adding support for [userfox](https://www.userfox.com)
+
+0.9.11 / 2013-4-26
+==================
+* Adding new ignoreReferrer option to Google Analytics provider
+* Adding new showFeedbackTab option to BugHerd provider
+* Updating UserVoice provider to work with their new snippet(s)
+* Fixing Errorception window.onerror binding to be friendlier
+
+0.9.10 / 2013-4-17
+=================
+* Adding url and title to mixpanel pageviews
+* Addiung url and title to keen pageviews
+
 0.9.9 / 2013-4-17
 =================
 * Fixed GoSquared relying on `document.body